+++ gems/aws-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased Changes
 ------------------
 
 * Feature - Asynchronously refresh AWS credentials.
+* Feature - Add support for recursion detection.
 
 3.125.6 (2022-02-02)
 ------------------